+++ app/routes/suggestions/index.js
@@ -4,6 +4,7 @@ var SuggestionsIndexRoute = Ember.Route.extend({
   model: function(params) {
     return this.modelFor('initiative').get('suggestions');
   }
+
 });
 
 export default SuggestionsIndexRoute;